Fix 3D-UNet to match MLCommons reference implementation
Preprocessing:
- Use fixed clip [-79.0, 304.0] and z-score mean=101.0, std=76.9
  (dataset-level, not per-volume) per nnU-Net global_vars.py
- Add resampling to target spacing [1.6, 1.2, 1.2] mm via scipy.ndimage.zoom
- Support MLPerf reference preprocessed pickle format (preprocessed_files.pkl)

Gaussian kernel:
- Use scipy.signal.gaussian with cube root transformation per reference
  inference_utils.py (sequential outer products -> g^(1/3) -> normalize)

Dice score:
- Use smooth_nr=smooth_dr=1e-6 (was 1e-5)
- Add smooth to numerator per reference accuracy_kits.py

Accuracy & LoadGen:
- Reference FP32 mean Dice is 0.86330, 99% threshold 0.85467
- Offline min_query_count: 43 per mlperf.conf (was 42)
- Use uint8 dtype for argmax output per reference

+++ src/mlperf_openvino/core/unet3d_sut.py
@@ -192,8 +192,8 @@ class UNet3DSUT:
         weight_sum = np.clip(weight_sum, a_min=1e-7, a_max=None)
         aggregated /= weight_sum[np.newaxis, ...]
 
-        # Argmax to get class predictions
-        segmentation = np.argmax(aggregated, axis=0).astype(np.int64)
+        # Argmax to get class predictions (uint8 per MLPerf reference)
+        segmentation = np.argmax(aggregated, axis=0).astype(np.uint8)
 
         return segmentation
